Replace dynamically size Vec with fixed-length array
Since an answer is always a static, 4-digit sequence, you can use a
fixed array and the length becomes part of the type. This afford some
safety because the compiler can check that you don't do something like
`answer[5]`, which would cause an index out of bounds panic

Additionaly, add a type alias called Answer so that it's easier to refer
to it than typing out the explici type `[u32; 4]` everywhere.